requirements.txt
pytest==7.1.2
pytest-xdist==2.5.0
requests==2.27.1
selenium==4.1.3
webdriver-manager==3.5.4
faker==13.11.1
allure-pytest==2.9.45
mysql-connector-python==8.0.22
